requirements.txt
tri.struct
tri.named_struct
django
pyparsing
tri.table
django-static-precompiler[libsass]